% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_colors.R
\name{add_colors}
\alias{add_colors}
\title{Add Colors to Environment and Display}
\usage{
add_colors(
  hex_codes = "#41536F",
  color_names = "BLUE",
  position = 3,
  colors_to_env = TRUE,
  show_colors = TRUE
)
}
\arguments{
\item{hex_codes}{A vector of hex color codes.}

\item{color_names}{A vector of color names.}

\item{position}{A numeric vector indicating the position of the color in the gradient.}

\item{colors_to_env}{Logical, whether to add colors to the environment.}

\item{show_colors}{Logical, whether to display the colors.}
}
\value{
None.
}
\description{
This function adds color palettes to the environment and optionally displays them.
}
