\name{trace_mat}
\alias{trace_mat}
\title{Compute the trace of a matrix}
\description{
  Computes the trace of a numeric matrix, which is the sum of its diagonal elements.
}
\usage{
  trace_mat(matrix_)
}
\arguments{
  \item{matrix_}{A numeric matrix.}
}
\value{
  A numeric value representing the trace of the matrix.
}
\author{
  Laval Jacquin, \email{jacquin.julien@gmail.com}
}
\examples{
  # Example usage
  # trace_value <- trace_mat(matrix_)
}
