\name{wiseR}
\alias{wiseR}
\title{wiseR: Helping your decisions from complex datasets}
\usage{
wiser()
}
\description{
wiseR is an interactive,end-to-end Bayesian Network Analysis and publishing platform for learning cross-validated structure, inferences, topology, visualization in interventional and observational data. It is powered by R, Shinydashboard, bnlearn and visNetwork packages.
}
\examples{
if (interactive()) {
  wiser()
}
}
\seealso{
\url{https://github.com/SAFE-ICU/wiseR}
}
