% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiser_datasets-documentation.R
\docType{data}
\name{wiser_datasets_description}
\alias{wiser_datasets_description}
\title{wiser datasets documentation}
\format{
rice_raw_pheno_data: a data frame with 384 rows and 15 columns.

rice_genomic_data: a data frame with 30 rows and 9929 columns.

maize_raw_pheno_data: a data frame with 2279 rows and 23 columns.

maize_genomic_data: a data frame with 30 rows and 41723 columns.

apple_raw_pheno_data: a data frame with 2117 rows and 24 columns.

apple_genomic_data: a data frame with 30 rows and 50001 columns.

pine_raw_pheno_data: a data frame with 180 rows and 16 columns.

pine_genomic_data: a data frame with 30 rows and 20796 columns.
}
\source{
https://academic.oup.com/g3journal/article/11/12/jkab320/6366549

https://pubmed.ncbi.nlm.nih.gov/31110353/

https://www.nature.com/articles/s41438-020-00408-8

https://pmc.ncbi.nlm.nih.gov/articles/PMC8867712/
}
\description{
These datasets are small subsets of 30 genotypes derived from the original datasets used
in Jacquin et al. (2025). They are provided for illustrative purposes with the wiser package
and are not intended to represent reference populations for genomic prediction or GWAS.
The datasets include genomic data and raw individual phenotypic measurements for the 30 selected
genotypes, aligned with the experimental design for each of the following species: rice, maize, apple, and pine.
}
